refactor: reorg app.py
